+++ data/headlines.json
@@ -1,0 +1,42 @@
+[
+  {
+      "title": "'집값 125번 조작' 文정부 고위 관료 11명 재판행…반대에 예산 삭감 압박",
+      "link": "https://n.news.naver.com/mnews/article/008/0005011986"
+  },
+  {
+      "title": "킬러문항·의대 입시에 사교육비 늘었나…지난해 27조 ‘역대 최대’ 찍어",
+      "link": "https://n.news.naver.com/mnews/article/081/0003437206"
+  },
+  {
+      "title": "\"꽃으로도 때리지 말라 했다\"…꽃 선물 사온 자녀 학대한 부부 꾸짖은 판사",
+      "link": "https://n.news.naver.com/mnews/article/666/0000036198"
+  },
+  {
+      "title": "의대 증원 비수도권 80%…'비수도권 미니 의대' 최대 수혜 전망",
+      "link": "https://n.news.naver.com/mnews/article/001/0014564129"
+  },
+  {
+      "title": "\"법정에 아기 계속 데려올거냐\"는 판사 질문에…이선균 협박범 \"네\"",
+      "link": "https://n.news.naver.com/mnews/article/277/0005392292"
+  },
+  {
+      "title": "출구 못 찾는 의정 갈등…의대교수 집단행동 움직임 가시화(종합)",
+      "link": "https://n.news.naver.com/mnews/article/001/0014564054"
+  },
+  {
+      "title": "\"담배 피러 간다며 도망다녀라\"…이번엔 '공보의 태업 지침'까지 등장",
+      "link": "https://n.news.naver.com/mnews/article/586/0000074549"
+  },
+  {
+      "title": "대통령실 \"이종섭 호주 대사 임명 철회 안해…공수처, 임명 전까지 왜 소환 조사 안했나?\"",
+      "link": "https://n.news.naver.com/mnews/article/087/0001031955"
+  },
+  {
+      "title": "\"유재석입니다\" 사칭에 1200억 털렸다…억울하게 불똥 튄 120명 분노",
+      "link": "https://n.news.naver.com/mnews/article/008/0005011960"
+  },
+  {
+      "title": "직원 얼굴에 유리 재떨이 던진 대표…항소심에선 감형받았다",
+      "link": "https://n.news.naver.com/mnews/article/081/0003437196"
+  }
+]
