--- headlines.json
@@ -1,12 +1,0 @@
-[
-  "전공의 4천944명 처분통지…절차 마무리 전 복귀하면 적극 선처",
-  "전공의 4천944명 처분통지…절차 마무리 전 복귀하면 적극 선처'탁구게이트' 이강인·손흥민 함께 태국전…주민규 첫 태극마크",
-  "'탁구게이트' 이강인·손흥민 함께 태국전…주민규 첫 태극마크野, '이종섭 호주 출국' 일제히 비난…\"개구멍으로 도망시켜\"",
-  "野, '이종섭 호주 출국' 일제히 비난…\"개구멍으로 도망시켜\"이변은 없었다…아카데미 작품상에 '오펜하이머'",
-  "이변은 없었다…아카데미 작품상에 '오펜하이머'與 선대위 '한동훈 원톱' 유력검토…'비례는 인요한 지휘' 거론",
-  "與 선대위 '한동훈 원톱' 유력검토…'비례는 인요한 지휘' 거론[속보] 尹 \"주민 원하는 곳에 케이블카 추가 건설…산림규제 대폭 완화\"",
-  "[속보] 尹 \"주민 원하는 곳에 케이블카 추가 건설…산림규제 대폭 완화\"'제약사 직원 동원 의혹' 고소인 조사…의협 \"동원지시 없었다\"",
-  "강원도민일보속보윤 대통령 \"오색케이블카 이어 강원도에 케이블카 추가 건설 계획\"",
-  "매일신문속보계양을 이재명 43%·원희룡 35%…분당갑 안철수 40%·이광재 36%",
-  "디지털타임스속보尹대통령 \"주민 원하는 곳에 케이블카 추가 건설…설악산 1300억 효과\""
-]
