+++ src/js/refreshPage.js
